Update gene values to floating point
Update gene values to be a list of floating point numbers instead of
nested tuples of ints of specific ranges. When we add more complicated
rendering this will allow us to be more dynamic. Eventually, a single
gene value could be used for different things depending on another
gene's value. This allows us to dynamically set the range.